requirements.txt
bcrypt==4.1.2
Flask==2.3.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Cors==3.0.10
psycopg2-binary
PyJWT==2.5.0
sqlalchemy
